Create a single JavaScript engine
